clean up formatting

+++ public/handlers.js
@@ -2,36 +2,35 @@
 
 //---------BUTTON CLICKS------------------
 //handles clicking on sign up button to display signup form
-function handleShowSignUp (){
+function handleShowSignUp() {
     $('body').on("click", "#js-show-sign-up-form-btn", (event) => {
         displaySignUpForm($('.js-content'));
     });
 }
 
 //handles clicking on log in button to display login form
-function handleShowLogIn (){
+function handleShowLogIn() {
     $('body').on("click", ".js-show-log-in-form-btn", (event) => {
         displayLogInForm($('.js-content'));
     });
 }
 
 //handles clicking on cancel button when adding new idea
-function handleCancelbtn(){
+function handleCancelbtn() {
     $('body').on("click", ".cancel-btn", (event) => {
         (user) ? getAndDisplayIdeas() : displayWelcomeHTML('.js-content');
-        });
+    });
 }
 
-
 //handles clicking on new idea button to display new idea form
-function handleShowNewIdeaForm(){
+function handleShowNewIdeaForm() {
     $('body').on("click", ".js-show-new-idea-form-btn", (event) => {
         displayNewIdeaForm($('.js-content'));
     });
 }
 
 //handles clicking on show idea details button 
-function handleShowIdeaDetails(){
+function handleShowIdeaDetails() {
     $('body').on("click", ".js-show-idea-details-btn", (event) => {
         const ideaID = $(event.currentTarget).data('id');
         const success = response => {
@@ -39,25 +38,24 @@ function handleShowIdeaDetails(){
         }
 
         getIdeaDetails({
-            jwToken: user, 
+            jwToken: user,
             ideaID,
             onSuccess: success
         });
     });
 }
 
-
 //handles clicking on edit idea button
-function handleShowEditableIdeaForm(){
+function handleShowEditableIdeaForm() {
     $('body').on("click", ".js-show-edit-idea-form-btn", (event) => {
-
         const ideaID = $(event.currentTarget).data('id');
         console.log(ideaID);
         const success = response => {
             displayEditableIdeaForm(response, ('.js-content'), false);
         }
+
         getIdeaDetails({
-            jwToken: user, 
+            jwToken: user,
             ideaID,
             onSuccess: success
         });
@@ -65,29 +63,36 @@ function handleShowEditableIdeaForm(){
     });
 }
 
-
 //handles deleting an idea after clicking on delete button
-function handleDeleteIdea(){
+function handleDeleteIdea() {
     $('body').on("click", ".js-delete-idea-btn", (event) => {
-        const ideaID = $(event.currentTarget).data('id');
-        console.log(ideaID);
-        deleteIdea({jwToken: user, ideaID, onSuccess: getAndDisplayIdeas, onError: ""});
+        var result = confirm("Are you sure you want to delete this idea?");
+        if (result) {
+            const ideaID = $(event.currentTarget).data('id');
+            deleteIdea({ jwToken: user, ideaID, onSuccess: getAndDisplayIdeas, onError: "" });
+        }
     });
 }
 
 //handles logging a user out
-function handleLogOut (){
+function handleLogOut() {
     $('body').on("click", ".js-log-out-btn", (event) => {
-        console.log("logout button")
         user = null;
         username = null;
         window.location.reload();
-});
+    });
+}
+
+//handles clicking on home icon in nav 
+function handleHomeIcon() {
+    $('body').on("click", ".js-home-btn", (event) => {
+        getAndDisplayIdeas();
+    });
 }
 
 //---------SUBMITTING FORMS------------------
 //handles submitting sign up/new user form
-function handleSignUpSubmit (){
+function handleSignUpSubmit() {
     $('body').on("submit", "#form-sign-up", (event) => {
         event.preventDefault();
         const newUserData = {
@@ -99,18 +104,18 @@ function handleSignUpSubmit (){
         }
 
         doNewUserCreation({
-            newUserData, 
+            newUserData,
             onSuccess: displaySignUpSuccessHTML('.js-content'),
             onError: err => {
                 alert('There was a problem accessing your request, please try again later.');
             }
         });
     });
-   
+
 }
 
 //handles submiting Log In form 
-function handleLogInSubmit (){
+function handleLogInSubmit() {
     $('body').on("submit", "#form-log-in", (event) => {
         event.preventDefault();
         const credentials = {
@@ -118,8 +123,8 @@ function handleLogInSubmit (){
             password: $('#password-txt').val()
         }
         doUserLogIn({
-            credentials, 
-            onSuccess: getAndDisplayIdeas, 
+            credentials,
+            onSuccess: getAndDisplayIdeas,
             onError: err => {
                 alert('Incorrect username or password. Please try again.');
             }
@@ -127,18 +132,15 @@ function handleLogInSubmit (){
     });
 }
 
-
-
 //handles submitting a new idea form/details
-function handleCreateNewIdea (){
+function handleCreateNewIdea() {
     $('body').on("submit", "#form-new-idea", (event) => {
         event.preventDefault();
-        const statusElement= document.getElementById('status-selector');
+        const statusElement = document.getElementById('status-selector');
         const statusVal = statusElement.options[statusElement.selectedIndex].value;
         const likabilityElement = document.getElementById('likability-selector');
         const likabilityVal = likabilityElement.options[likabilityElement.selectedIndex].value;
-        
-        
+
         const newIdea = {
             title: $('#title-txt').val(),
             description: document.getElementById("description-txt").value,
@@ -147,20 +149,20 @@ function handleCreateNewIdea (){
         }
 
         createNewIdea({
-            jwToken: user, 
-            newIdea, 
+            jwToken: user,
+            newIdea,
             onSuccess: getAndDisplayIdeas,
             onError: ""
         });
     });
 }
 
-//handles submitting update idea form/details
-function handleUpdateIdea (){
+//handles submitting update idea form
+function handleUpdateIdea() {
     $('body').on("submit", "#form-update-idea", (event) => {
         event.preventDefault();
         const ideaID = $(event.currentTarget).data('id');
-        const statusElement= document.getElementById('status-selector');
+        const statusElement = document.getElementById('status-selector');
         const statusVal = statusElement.options[statusElement.selectedIndex].value;
         const likabilityElement = document.getElementById('likability-selector');
         const likabilityVal = likabilityElement.options[likabilityElement.selectedIndex].value;
@@ -171,20 +173,19 @@ function handleUpdateIdea (){
             status: statusVal,
             likability: likabilityVal
         }
-       
+
         updateIdea({
-            jwToken: user, 
+            jwToken: user,
             ideaID,
             updatedIdea,
-            onSuccess: getAndDisplayIdeas, 
+            onSuccess: getAndDisplayIdeas,
             onError: ""
         });
     });
 }
-//add error handling
 
-
-function setUpEventHandlers(){
+//handles setting up all eventhandlers
+function setUpEventHandlers() {
     handleShowSignUp();
     handleSignUpSubmit();
     handleShowLogIn();
@@ -196,5 +197,6 @@ function setUpEventHandlers(){
     handleUpdateIdea();
     handleDeleteIdea();
     handleShowNewIdeaForm();
-    handleCancelbtn ()
+    handleCancelbtn();
+    handleHomeIcon();
 }
